Fixed all warnings on MSVC, fixed up usage of int instead of size_t, fixed some signatures.

+++ spine-c/spine-c/src/spine/Json.c
@@ -147,7 +147,7 @@ static const char* parse_number (Json *item, const char* num) {
 
 	if (ptr != num) {
 		/* Parse success, number found. */
-		item->valueFloat = result;
+		item->valueFloat = (float)result;
 		item->valueInt = (int)result;
 		item->type = Json_Number;
 		return ptr;
