+++ spine-c/spine-c/src/spine/SkeletonJson.c
@@ -93,7 +93,7 @@ static float toColor (const char* value, int index) {
 	char *error;
 	int color;
 
-	if (index >= strlen(value) / 2)
+	if ((size_t)index >= strlen(value) / 2)
 		return -1;
 	value += index * 2;
