+++ spine-cpp/spine-cpp/include/spine/VertexAttachment.h
@@ -59,7 +59,7 @@ namespace Spine {
         /// @param worldVertices The output world vertices. Must have a length greater than or equal to offset + count.
         /// @param offset The worldVertices index to begin writing values.
         /// @param stride The number of worldVertices entries between the value pairs written.
-        void computeWorldVertices(Slot& slot, int start, int count, Vector<float>& worldVertices, int offset, int stride = 2);
+        void computeWorldVertices(Slot& slot, size_t start, size_t count, Vector<float>& worldVertices, size_t offset, size_t stride = 2);
         
         /// @return true if a deform originally applied to the specified attachment should be applied to this attachment.
         virtual bool applyDeform(VertexAttachment* sourceAttachment);
@@ -67,17 +67,17 @@ namespace Spine {
         /// Gets a unique ID for this attachment.
         int getId();
         
-        Vector<int>& getBones();
+        Vector<size_t>& getBones();
         
         Vector<float>& getVertices();
         
-        int getWorldVerticesLength();
-        void setWorldVerticesLength(int inValue);
+        size_t getWorldVerticesLength();
+        void setWorldVerticesLength(size_t inValue);
         
     protected:
-        Vector<int> _bones;
+        Vector<size_t> _bones;
         Vector<float> _vertices;
-        int _worldVerticesLength;
+        size_t _worldVerticesLength;
         
     private:
         const int _id;
