+++ spine-cpp/spine-cpp/src/spine/LinkedMesh.cpp
@@ -34,7 +34,7 @@
 
 using namespace Spine;
 
-LinkedMesh::LinkedMesh(MeshAttachment *mesh, const String &skin, int slotIndex, const String &parent) :
+LinkedMesh::LinkedMesh(MeshAttachment *mesh, const String &skin, size_t slotIndex, const String &parent) :
 		_mesh(mesh),
 		_skin(skin),
 		_slotIndex(slotIndex),
